pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[anthropic,google-genai,openai]>=1.0.5",
    "python-dotenv>=1.2.1",
]
